--- a/hw10_program_optimization/stats.go
+++ b/hw10_program_optimization/stats.go
@@ -6,7 +6,7 @@ import (
 	"io"
 	"strings"
 
-	"github.com/json-iterator/go"
+	jsoniter "github.com/json-iterator/go"
 )
 
 type User struct {